Dashboard Completed
I updated the filter panel and added the charts. The charts are displayed in 2 sections.

The first section displays a Line and a clustered chart. Column bars show total sales, and the line shows the average based on the selected parameter over time.

The 2nd section displays two charts. The first chart shows the trend in total sales over time. The 2nd chart is a tree-map based on the total sales of each store.